services/results_api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
redis
websockets
starlette-prometheus